some changes 14-37 11/10/2021
