+++ src/test/java/Pages/SendMailPage.java
@@ -1,4 +1,4 @@
-package Config;
+package Pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
